SLING-7318 no longer rely on commons-osgi for the test
Commons OSGi was referred to in an old version (which was not
necessarily contained in the m2 repo). Commons OSGi is is not a
test-only dependency and therefore the version might change in the
future. For the test only dependencies it is easier to ensure that those
versions don't change.